Created a new loading screen
