--- todo-list-client/src/components/shared/themeButton/ThemeButton.enums.ts
@@ -1,4 +1,0 @@
-export enum ThemesEnum {
-  LIGHT = 'light',
-  DARK = 'dark',
-}
