+++ todo-list-client/src/design/base/_global.scss
@@ -15,6 +15,7 @@ main {
   flex-direction: column;
   justify-content: center;
   align-items: center;
+  background-color: var(--neutral-one);
 }
 
 a {
