+++ todo-list-client/src/enums/theme.ts
@@ -1,0 +1,4 @@
+export enum ThemesEnum {
+    LIGHT = 'light',
+    DARK = 'dark',
+}
