+++ todo-list-client/src/router/Router.component.tsx
@@ -2,7 +2,11 @@ import { Routes } from 'react-router-dom';
 import { ApplicationRoutes } from './Router.blocks';
 
 const Router = () => {
-  return <Routes>{ApplicationRoutes}</Routes>;
+  return (
+    <Routes>
+      {ApplicationRoutes}
+    </Routes>
+  );
 };
 
 export default Router;
