+++ todo-list-client/src/router/Router.constants.tsx
@@ -3,10 +3,13 @@ import i18n from 'i18next';
 import Auth from '../screens/Auth';
 import Home from '../screens/Home';
 import List from '../screens/List';
+import { LanguagesEnum } from '../components/shared/localeButton/LocaleButton.enums';
 
 export const APPLICATION_ROUTES: Array<RouteProps> = [
   { path: '/', element: <Navigate to={`/${i18n.language}/auth`} replace /> },
-  { path: '/:lang/auth', Component: Auth },
-  { path: '/:lang/home', Component: Home },
-  { path: '/:lang/lists/:listId', Component: List },
+  ...Object.values(LanguagesEnum).map((lang) => [
+    { path: `/${lang}/auth`, element: <Auth /> },
+    { path: `/${lang}/home`, element: <Home /> },
+    { path: `/${lang}/lists/:listId`, element: <List /> },
+  ]).flat(),
 ];
